Application structure and some renaming
