+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/model/SecurityClassification.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.roleassignment.model.enums;
+package uk.gov.hmcts.reform.roleassignment.domain.model;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
