+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/common/RetrieveSupportingDataService.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.common;
+
+public class RetrieveSupportingDataService {
+}
