+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/common/StoreRequestService.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.common;
+
+public class StoreRequestService {
+}
