+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/common/UpdateStatusService.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.common;
+
+public class UpdateStatusService {
+}
