+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/createresources/CreateRoleAssignments.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.createresources;
+
+public class CreateRoleAssignments {
+}
