+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/validateresource/ValidateResourceOperation.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.validateresource;
+
+public class ValidateResourceOperation {
+}
