+++ src/test/java/uk/gov/hmcts/reform/roleassignment/model/enums/SecurityClassificationTest.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.reform.roleassignment.model.enums;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import org.junit.jupiter.api.Test;
+import uk.gov.hmcts.reform.roleassignment.domain.model.SecurityClassification;
 
 class SecurityClassificationTest {
